src/layout/UserAuth.js: init, layout template for Register and Login pages
